+++ src/form/field/CheckboxField.tsx
@@ -3,7 +3,9 @@ import { Checkbox } from "../../control/Checkbox";
 import { Field } from "formik";
 
 export interface ICheckboxFieldProps {
+    id?: string;
     name: string;
+    value?: string;
     required?: boolean;
     disabled?: boolean;
 }