+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# Unreleased
+
+- (form): Add CustomField component as a formik Field wrapper
+- (CheckboxField, InputField, RadioField, TextareaField): Forward `id` attribute
+- (CheckboxField): Forward `value` attribute (for checkbox arrays)
+- (SelectField): Fix validation not triggering when selecting an option
+
 # v2.0.0-dev.16
 
 - Add new icons: IdentityIcon, MailIcon, SlackIcon, TriangleDownIcon, TriangleUpIcon, TriangleLeftIcon, TriangleRightIcon